♻️ Split component
