+++ src/component/Header.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <header>
+      <h1>React-Redux 掲示板</h1>
+    </header>
+  );
+}
+
+export default Header
