+++ src/component/Main.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import PostList from "component/PostList";
+
+const Main = () => {
+  return (
+    <main>
+      <PostList />
+    </main>
+  );
+};
+
+export default Main;
